python-ai/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1

# Database clients
neo4j==5.17.0
redis==5.0.2
weaviate-client>=4.5.0

# gRPC
grpcio==1.62.0
grpcio-tools==1.62.0
grpcio-reflection==1.62.0
grpcio-health-checking==1.62.0

# AI/ML
numpy>=1.26.0,<1.27.0
torch>=2.2.0
transformers>=4.36.0
sentence-transformers>=2.3.0
scikit-learn>=1.4.0
nltk>=3.8.1

# AI API clients
openai>=1.12.0
anthropic>=0.8.1
mistralai>=0.0.8

# Content processing
beautifulsoup4>=4.12.3
lxml>=5.1.0
tqdm>=4.66.2

# Development tools
pytest==7.4.4
pytest-cov==4.1.0
black==24.2.0
pylint==3.0.3
debugpy==1.8.0

# Utilities
requests==2.31.0
httpx==0.27.0
tenacity==8.2.3